LayoutTests/storage/indexeddb/modern/versionchange-event-expected.txt

This test: -Opens a connection to a database at version 1, creating the database -Commits the version change transaction for that database -Opens a second connection to that database, requesting version 1 -Opens a third connection to that database, requesting version 2 -Makes sure the first and second connections get the versionchange event -Closes the first and second connections -Makes sure the versionchange transaction for the second connection starts successfully

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: upgradeneeded (firstPhase): old version - 0 new version - 1
ALERT: Version change complete (firstPhase). Database version is now - 1
ALERT: First version change successful
ALERT: Open success (secondPhase)
ALERT: thirdPhase - Requested database connection with version 2
ALERT: Expected upgrade needed (thirdPhase)
ALERT: firstPhase connection had received oldVersion: 1, newVersion: 2
ALERT: secondPhase connection had received oldVersion: 1, newVersion: 2
PASS successfullyParsed is true

TEST COMPLETE
